refactor: drop owner in favor of ethaddress
